01-todo/requirements-dev.txt
coverage
pytest
pytest-django
pytest-cov
factory-boy
pytest-factoryboy
freezegun
ruff
